TI: remoteproc: k3-r5: Move suspend to suspend_late
The remote core has to be able to query the DM for the next system mode
in the suspend path. To support this, ti_sci.c has to send the
prepare_sleep command before the suspend message is sent to the remote
core. This patch moves the suspend to suspend_late to be executed after
ti_sci's suspend call.

Change-Id: I784968e71030f9216cc0fa7a99ef3db25fa5d36d
Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
(cherry picked from commit 6b5a010ed2c4b727e3a487690c11b0c652190872)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>