TI: remoteproc: k3-m4: Move suspend to suspend_late
The remote core has to be able to query the DM for the next system mode
in the suspend path. To support this, ti_sci.c has to send the
prepare_sleep command before the suspend message is sent to the remote
core. This patch moves the suspend to suspend_late to be executed after
ti_sci's suspend call.

Change-Id: I425c36ecc67c05def73f006b767b8af951e52d41
Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
(cherry picked from commit 49b236a4e51e804a1363289c6bdb6d7fb1037e78)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>