TI: arm64: dts: ti: k3-am62a7-sk: HACK: Reduce mailbox fifos
As such each mailbox cluster in AM62A have 16 fifos. For Linux rpmsg
IPC only 2 fifos are used. The rest of the fifos could be used for
RTOS-RTOS communications.

On a suspend, mailbox driver checks for any pending outstanding
messages in the fifos and if any present returns with an error.
Currently C7x DSP is not under system suspend and it seems some
mailbox messages between DM-R5 and C7 could be present in the
fifos when Linux suspend is exercised. This is resulting in mailbox
driver suspend failure. As a HACK for now restrict the mailbox fifos
to 2, which are used for Linux IPC with remote core.

Change-Id: Ibf28bab569fa1f0571922bfc77b7227de32df353
Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
(cherry picked from commit 4c9b145cd486028900954681de70d3d8fd137d1f)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>