TI: media: chips-media: wave5: Add support for dmabuf import
Wave5 decoder does not guarantee that the order of frames being queued to
the driver matches the order of frames being displayed. This behavior is
acceptable as there is nowhere is spec saying order needs to match.
However, this leads to issues in dmabuf import mode  where userspace does
not maintain mapping of video buffers (vb2_buffer) queued to driver and the
physical address it is assigned to. This patch adds support for maintaining
that mapping such that downstream buffers can be used for zero copy.

Change-Id: I909d282cedbbbfb3d51261f7f3872601b05bbc37
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Tested-by: Rahul T R <r-ravikumar@ti.com>
Reviewed-by: Darren Etheridge <detheridge@ti.com>
(cherry picked from commit 1d17f0e0791113f27e3c018cf4903dad448b865d)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>