TI: net: ti: icssg_prueth: Enable 10M Link issue quirk for AM64x
There is a HW errata for 10M issues which is present on AM65x but not on
AM64x. The workaround for this errata is to enable IEP1. Since both
AM64x and AM65x shares the same ICSSG firmware, firmware is expecting
IEP1 to be enabled for both AM64x and AM65x. As a result 10M Link fails
on AM64x.

Enable the quirk_10m_link_issue so that IEP1 can be enabled for AM64x

Change-Id: I0651828709048b92fe6eecd4f543be2eac85a765
Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
Reviewed-by: Daolin Qiu <d-qiu@ti.com>
(cherry picked from commit 6de6e418c80edfbe08f4a5f851c721bd60c0123b)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>