UPSTREAM: tools/lib/thermal: Add the threshold netlink ABI
The thermal framework supports the thresholds and allows the userspace
to create, delete, flush, get the list of the thresholds as well as
getting the list of the thresholds set for a specific thermal zone.

Add the netlink abstraction in the thermal library to take full
advantage of thresholds for the userspace program.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Reviewed-by: Lukasz Luba <lukasz.luba@arm.com>
Link: https://patch.msgid.link/20241022155147.463475-5-daniel.lezcano@linaro.org
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Bug: 390217395
Change-Id: I19381425728a3ff91690276dee78f08f2ba6d28c
(cherry picked from commit a26267248628760d02e7d03534afc59dae0c9b40)
Signed-off-by: Manaf Meethalavalappu Pallikunhi <quic_manafm@quicinc.com>