TI: drm/tidss: Remove extra K2G check
We check if the platform is K2G in dispc_k3_clear_irqstatus(), and
return early if so. This cannot happen, as the _k3_ functions are never
called on K2G in the first place. So remove the check.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Swamil Jain <s-jain1@ti.com>
(cherry picked from commit d83298d251401ad6483d1475ac79cd994a27ba6a)
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>