ANDROID: GKI: rexport __qdisc_calculate_pkt_len
The symbol __qdisc_calculate_pkt_len was removed from the .h file in
1edf039ee017 ("net/sched: accept TCA_STAB only for root qdisc") but it
still lives on in some export symbol lists.  It's not really needed if
they rebuild, but for now, let's just keep it exported so that
everything continues to build properly, as it's safe to do it this way.

Fixes: 1edf039ee017 ("net/sched: accept TCA_STAB only for root qdisc")
Change-Id: If549f5074d022b6f0d0d2ab7a566eeb6e7db458a
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>