BACKPORT: TI: remoteproc: k3-dsp: Introduce PM suspend/resume handlers
Introduce suspend/resume handling to dsp remoteproc driver. The driver
registers for pm notifications. On a linux suspend a mailbox message
'RP_MBOX_SUSPEND_SYSTEM' is sent to the remote core.
a) If the remote core doesn't respond to the SUSPEND message, the
driver sets a device ON constraint and returns.
b) If the remote core responds with 'RP_MBOX_SUSPEND_ACK', the driver
initiates a stop of the remote core and returns from the suspend
handler.
c) On the other hand, if the remote core responds with
RP_MBOX_SUSPEND_AUTO the driver simply returns.

On the Resume path, driver queries DM to determine the state of the
remote core and if it is OFF, turns ON the remote core.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com>
(cherry picked from commit 2d1562b82f9f44a6261a8f5c0720285a650a7a07)

Conflicts:
   drivers/remoteproc/ti_k3_dsp_remoteproc.c
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>