HACK: ARM: OMAP2+: powerdomain: Do not register pwrdms for early booted rprocs
In case of Early Boot of IPU1, the rproc has already been powered-on and
booted by the bootloader. Therefore, as a HACK, do not register the
pwrdm data for IPU1 in Kernel, to prevent Kernel from resetting it at
init time.

The above limitation exists because IPUs (and other rprocs) have not
been migrated to be used by the ti-sysc bus driver, which utilizes the
DT quirks like "ti,no-reset-on-init" prevent resets.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>