HACK: ARM: OMAP2+: clockdomain: Do not initialize clks for early booted rprocs
In case of Early Boot of IPU1, the rproc has already been powered-on and
booted by the bootloader. Therefore, as a HACK, do not initialize the
clks for IPU1 in Kernel, to prevent Kernel from configuring clks at init
time.

The clkdata for IPU1 is mostly present in "clockdomains7xx_data.c" file,
and few are also hardcoded in the "clk-7xx.c" driver. Take care of all
the instances by skipping the initialization of those clkdatas. The
above limitation exists because IPUs (and other rprocs) have not been
migrated to be used by the ti-sysc bus driver.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>