HACK: dma: coherent: Do not clear allocated buffer
The default behaviour of the "__dma_alloc_from_coherent()" function is
to clear the allocated buffer. In case of Early Boot, the firmware
already resides in the carveout locations, and allocation is only done
to reserve that memory from Linux point of view. Therefore, avoid
clearing the memory upon allocation for early booted remoteprocs.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>