HACK: remoteproc: omap: add "late attach" support
Add the necessary support in the OMAP remoteproc driver to enable the
"late attach" functionality. The 'late attach' functionality has the
processors already loaded and running by the time the kernel has
booted.

The 'late attach' support in OMAP remoteproc driver is added through
minimal changes like skipping the releasing of the processor reset.
The processor reset is skipped as this relies on the omap_hwmod /
omap_device API which perform a module disable/enable sequence in
addition to the reset programming. Other required functionality is
achieved through the late attach support in the remoteproc driver core.

The 'late attach' capability is determined through a DT-property called
'late_attach'. The rproc nodes along with the corresponding IOMMU and
timer nodes should all have the "ti,no-idle-on-init" and
"ti,no-reset-on-init" properties set to prevent omap_hwmod from
resetting and idling these devices on boot. The .late_attach field in
the rproc data structure is used to denote the mode and differentiate a
normal boot from a late-attach boot.

Since the rproc IP blocks have not been migrated to be used by the
ti-sysc driver, as a HACK, the pwrdms and clkdms for these rprocs have
not been registered at Kernel init time. As a result, the runtime
suspend and resume operations can not be utilized for these rprocs.

Signed-off-by: Robert Tivy <rtivy@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Shravan Karthik <shravan.karthik@ti.com>
Signed-off-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>