net: ti: icssg-prueth: Drop duplicate definition of icssg_read_time()
The fix [0] is already present in the ti-linux-6.6.y-cicd tree. During
stable backport by [1], the counterpart of [0] on the stable tree [2] was
also introduced in ti-linux-6.6.y-cicd, resulting in the duplicate
definition of "icssg_read_time()" in "icssg_prueth.h". Since the stable
backport is not a perfect duplicate of [0] (changes in "icssg_prueth.c"
aren't duplicated), a revert isn't possible. Hence, fix this by manually
dropping the duplicate definition.

[0]: 6d5d39a59956 ("net: ti: icssg-prueth: Fix 1 PPS sync")
[1]: e68f4937d87d ("Merge tag 'v6.6.63' of git://git.kernel.org/pub/scm/linux/kernel/git/stable/linux into ti-linux-6.6.y-cicd")
[2]: c67ce71d5328 ("net: ti: icssg-prueth: Fix 1 PPS sync")
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>