UPSTREAM: thermal/thresholds: Fix uapi header macros leading to a compilation error
The macros giving the direction of the crossing thresholds use the BIT
macro which is not exported to the userspace. Consequently when an
userspace program includes the header, it fails to compile.

Replace the macros by their litteral to allow the compilation of
userspace program using this header.

Fixes: 445936f9e258 ("thermal: core: Add user thresholds support")
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Link: https://patch.msgid.link/20241212201311.4143196-1-daniel.lezcano@linaro.org
[ rjw: Add Fixes: ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Bug: 390217395
Change-Id: If7393a08b07de0af3297d728141174a26f64c989
(cherry picked from commit 65c8c78cc74d5bcbc43f1f785a004796a2d78360)
Signed-off-by: Manaf Meethalavalappu Pallikunhi <quic_manafm@quicinc.com>