FROMLIST: usb: host: xhci-mem: Cleanup pending secondary event ring events
As part of xHCI bus suspend, the xHCI is halted.  However, if there are
pending events in the secondary event ring, it is observed that the xHCI
controller stops responding to further commands upon host or device
initiated bus resume.  Iterate through all pending events and update the
dequeue pointer to the beginning of the event ring.

Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>

Bug: 392112015
Link: https://lore.kernel.org/linux-usb/20250121210518.2436771-3-quic_wcheng@quicinc.com/
Change-Id: I1ff5ef47ed91ea472f5fecf4aa4cdd610e2804d2
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>