UPSTREAM: regulator: tps65219: Remove debugging helper function
commit 64a6b577490c1c71f1a3bbdb3844717815214621 upstream.

This helper function is primarily used by developers for debugging & is
not a standard feature included in other PMIC drivers. The purpose of
debugging function is to check if rdev is determine if the regulator has
been registered prior to requesting a threaded irq. This case is already
handled with the ISERR(rdev) check because the error code is returned.

Link: https://lore.kernel.org/lkml/20241217204526.1010989-5-s-ramamoorthy@ti.com/
Signed-off-by: Shree Ramamoorthy <s-ramamoorthy@ti.com>
(cherry picked from commit cb5ee4ce3a68b6a55e79c6c8ebb4a20c23312719)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>