FROMLIST: remoteproc: k3: Refactor rproc_release() implementation into common driver
The k3_dsp_rproc_release() function erroneously deasserts the local
reset even for devices which do not support it. Even though it results
in a no-operation, Update the logic to explicitly deassert the local
reset for devices that support it and only the global reset for those
that do not.

Further, refactor the above function into the ti_k3_common.c driver as
k3_rproc_release() and use it throughout DSP and M4 drivers for
releasing the reset from the remote processor.

Link: https://lore.kernel.org/all/20250103101231.1508151-8-b-padhi@ti.com/
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
(cherry picked from commit 00158db3633b9d7be37e4fe6940a0548d7de0a46)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>