FROMLIST: drm/tidss: Add dispc_is_idle()
Add a helper function, dispc_is_idle(), which returns whether the DSS is
idle (i.e. is any video port enabled).

For now we add a call to it in the suspend and resume callbacks, and
print a warning if in either place the DSS is not idle.

Link: https://lore.kernel.org/all/20231101-tidss-probe-v1-8-45149e0f9415@ideasonboard.com/
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
(cherry picked from commit 43385cc26f163b4df605bc6ebfcdaaa7b0f0bd25)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>