FROMLIST: drm/tidss: Remove extra K2G check
We check if the platform is K2G in dispc_k3_clear_irqstatus(), and
return early if so. This cannot happen, as the _k3_ functions are never
called on K2G in the first place. So remove the check.

Link: https://lore.kernel.org/all/20241021-tidss-irq-fix-v1-3-82ddaec94e4a@ideasonboard.com/
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Swamil Jain <s-jain1@ti.com>
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
(cherry picked from commit 50f644f02e3c7c746eae5749685df585fadebf4f)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>