ANDROID: KVM: arm64: Calculate traps for all vcpus at hyp vm creation
In non-protected mode, vcpu traps are calculated for each vcpu
the first time it's run. Protected mode creates the hyp view of
the vm and all its vcpus when the first vcpu is run. At that point
in time, the remaining vcpus might not have calculated their trap
values.

To fix this problem, calculate vcpu traps for _all_ vcpus before
creating and initializing the hyp view.

Bug: 394097954
Bug: 357781595
Signed-off-by: Fuad Tabba <tabba@google.com>
Change-Id: I9fdccd035aa2d95775704b8e3dcbd6f549ae3c4c