UPSTREAM: cpufreq: ti-cpufreq: Allow backward compatibility for efuse syscon
commit 1724ae88efcbcd0daeb203ffeb4a2c0e59f2ddf7 upstream

The AM625 syscon for efuse was being taken earlier from the wkup_conf node
where the entire wkup_conf was marked as "syscon". This is wrong and will
be fixed in the devicetree. However, whenever that does happen will end up
breaking this driver for that device because of the change in efuse offset.

Hence, to avoid breaking any sort of backward compatibility of devicetrees
use a quirk to distinguish and accordingly use 0x0 offset for the new
syscon node.

Suggested-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>
Reviewed-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
(cherry picked from commit ae6546bf1b5cc875c51986fd79a57ad7c55e6206)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>