FROMLIST: firmware: ti_sci: Support transfers without response
Check the header flags if an response is expected or not. If it is not
expected skip the receive part of ti_sci_do_xfer(). This prepares the
driver for one-way messages as prepare_sleep for Partial-IO.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Link: https://lore.kernel.org/r/20241219-topic-am62-partialio-v6-12-b4-v4-1-1cb8eabd407e@baylibre.com
Signed-off-by: Kendall Willis <k-willis@ti.com>
(cherry picked from commit 26000aa62b0b77a6fa770dd2c05dcb26fbf85f8d)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>