FROMLIST: net: ti: icssg-prueth: introduce and use prueth_swdata struct for SWDATA
We have different cases for SWDATA (skb, page, cmd, etc)
so it is better to have a dedicated data structure for that.
We can embed the type field inside the struct and use it
to interpret the data in completion handlers.

Link: https://lore.kernel.org/all/20250305101422.1908370-3-m-malladi@ti.com/
Signed-off-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Meghana Malladi <m-malladi@ti.com>
(cherry picked from commit d7447f891c9b083a252e96c6849b55fc5d3bab88)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>