PENDING: remoteproc: k3: Add support for suspend/resume handlers
Introduce suspend/resume handling to k3 remoteproc driver. k3 remoteproc
drivers registers for pm notifications. On a linux suspend a mailbox message
'RP_MBOX_SUSPEND_SYSTEM' is sent to the remote core.
a) If the remote core doesn't respond to the SUSPEND message, the
driver sets a device ON constraint and returns.
b) If the remote core responds with 'RP_MBOX_SUSPEND_ACK', the driver
initiates a stop of the remote core and returns from the suspend
handler.
c) If the remote core responds with 'RP_MBOX_SUSPEND_CANCEL', the driver
returns with a -EBUSY error.
c) On the other hand, if the remote core responds with
RP_MBOX_SUSPEND_AUTO the driver simply returns, meaning the remote core
is handling the suspend by itself.

On the Resume path, driver queries DM to determine the state of the
remote core and if it is OFF, turns ON the remote core.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
(cherry picked from commit 83f58333b5e3e7529ad2e7923c33f12f0bddd019)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>