PENDING: drivers: crypto: sa2ul: Keep ahash as the only fallback transform
ahash is the only one that is used for all the fallback calls; This
allows us to use shash as a helper transform that can be allocated for
any other hash calculations that we need to do.

Split ahash and shash to allow HMAC to use the shash for computing the
intermediate digest.

Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
(cherry picked from commit b259dcfc3724348211530021968b17753f965da6)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>