Revert "ANDROID: enable memory allocation profiling configs"
This reverts commit 291f8636115d96a9ba20ce6142fad37599b0bcb9.
We see below compilation issues with our debug + kasan enabled builds,
revert the change as it is more of a debug feature.

Traceback (most recent call last): File "avbtool.py", line 3505, in add_hash_footer AvbError: Image size of 67772416 exceeds maximum image size of 67039232 in order to fit in a partition size of 67108864.

Bug: 402596474
Change-Id: Ie2f4dc97e60454fac44bc2f75527a3760cd68e60
Signed-off-by: Satya Durga Srinivasu Prabhala <quic_satyap@quicinc.com>