ANDROID: GKI: Enable CONFIG_BT_HCIBTUSB_AUTO_ISOC_ALT
This config is necessary for the HFP (bi-direction audio) on USB
Bluetooth controller + HCI User Channel Socket.

Since the btusb module was not compiled in ACK before, this means the
module was not used by any other projects and it shall be safe to enable
the config on ACK by default.

Bug: 387425332
Test: Verify HFP works on USB Bluetooth controller
Change-Id: Ib03c38f78929dbd4fbf335e2fe16d5627a211b9b
Signed-off-by: Hsin-chen Chuang <chharry@google.com>