ANDROID: irqchip/gic-v3-its: Export suspend and resume functions for S2IDLE
Although the suspend and resume functions exist and are registered as syscore_ops
in the upstream GICv3 ITS driver,
there is currently no proper way to invoke these functions in the S2IDLE scenario,
as the S2IDLE framework does not call syscore_ops.

To address this, export the suspend and resume symbols from the GICv3 ITS driver.
This allows vendor modules to hook these functions into the s2idle_ops callbacks,
enabling the save and restore of the GICv3 ITS state during S2IDLE.

Bug: 405256435
Signed-off-by: Youngmin Nam <youngmin.nam@samsung.com>
Change-Id: Ib91e722856d5034da771c73612b5ca9f002a3342