ANDROID: vendor_hooks: add hook to record binder transaction info
Add vendor hook to record binder transaction information and
support oem's print binder transaction info.

Our function is to help locate whether the problem is caused by binder
blocking when ANR or watchdog occurs;So we need to add timestamps,
synchronous or asynchronous process information to the binder_procs
node, and create our own file node to readbinder_procs,
binder_transaction_log and binder_transaction_log_failed,
and output them in the format we need.

On kernel 6.6, we implement this with an OGKI patch, now we want
to achieve this by adding this vendor hook (also helps to reduce
our OGKI patches).This is a custom function, which will affect
performance. I turn this unction on or off according to different
products. It is not available for all OEMs, so it is not available
upstream.

Bug: 409484697
Change-Id: I36109e9045e284cc8be8dbffb79fc5b509a67c06
Signed-off-by: zhengwei <zhengwei2@honor.com>