PENDING: arm64: dts: ti: k3-am62d2-evm: Update PHY Address for E2
The Ethernet PHY corresponding to CPSW3G's MAC Port1 is present at a PHY
Address of "15" in the E1 revision of the board. In the E2 revision of the
board, the PHY Address has changed to "0". Hence, update it.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
(cherry picked from commit 013211ccdf5ec33e19652be631b820bc5f6b86a4)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>