UPSTREAM: arm64: dts: ti: k3-am64-main: Switch to 64-bit address space for PCIe0
commit 1159f911435b9371575d9956bb5aef579265d29a upstream.

The PCIe0 instance of PCIe in AM64 SoC supports:
1. 128 MB address region in the 32-bit address space
2. 4 GB address region in the 64-bit address space

The default configuration is that of a 128 MB address region in the
32-bit address space. While this might be sufficient for most use-cases,
it is insufficient for supporting use-cases which require larger address
spaces. Therefore, switch to using the 64-bit address space with a 4 GB
address region.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Link: https://lore.kernel.org/r/20250422120042.3746004-2-s-vadapalli@ti.com
Signed-off-by: Nishanth Menon <nm@ti.com>
(cherry picked from commit 1601f3254db2284e7095d78f74bfebc38036592c)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>