UPSTREAM: arm64: dts: ti: k3-j7200-main: Switch to 64-bit address space for PCIe1
commit 46e3d7d7048872ffb90aa25e360b975dc1c3e28b upstream.

The PCIe0 instance of PCIe in J7200 SoC supports:
1. 128 MB address region in the 32-bit address space
2. 4 GB address region in the 64-bit address space

The default configuration is that of a 128 MB address region in the
32-bit address space. While this might be sufficient for most use-cases,
it is insufficient for supporting use-cases which require larger address
spaces. Therefore, switch to using the 64-bit address space with a 4 GB
address region.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Link: https://lore.kernel.org/r/20250422120042.3746004-3-s-vadapalli@ti.com
Signed-off-by: Nishanth Menon <nm@ti.com>
(cherry picked from commit a03bd4ea0b0e6714f84dabf4ebb360c890c1de36)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>