UPSTREAM: arm64: dts: ti: k3-j721e-main: Switch to 64-bit address space for PCIe0 and PCIe1
commit f0f78192d3b3b3c3c935de0d681e7bf2117bbb13 upstream.

The PCIe0 and PCIe1 instances of PCIe in J721E SoC support:
1. 128 MB address region in the 32-bit address space
2. 4 GB address region in the 64-bit address space

The default configuration is that of a 128 MB address region in the
32-bit address space. While this might be sufficient for most use-cases,
it is insufficient for supporting use-cases which require larger address
spaces. Therefore, switch to using the 64-bit address space with a 4 GB
address region.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Link: https://lore.kernel.org/r/20250422120042.3746004-5-s-vadapalli@ti.com
Signed-off-by: Nishanth Menon <nm@ti.com>
(cherry picked from commit b790488078fd6306d38b75265aba0a30faf36920)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>