UPSTREAM: arm64: dts: ti: k3-j721s2-main: Switch to 64-bit address space for PCIe1
commit 5a765365c689b90ea81f651a43bc30cd9c9c4ed8 upstream.

The PCIe1 instance of PCIe in J721S2 SoC supports:
1. 128 MB address region in the 32-bit address space
2. 4 GB address region in the 64-bit address space

The default configuration is that of a 128 MB address region in the
32-bit address space. While this might be sufficient for most use-cases,
it is insufficient for supporting use-cases which require larger address
spaces. Therefore, switch to using the 64-bit address space with a 4 GB
address region.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Link: https://lore.kernel.org/r/20250422120042.3746004-6-s-vadapalli@ti.com
Signed-off-by: Nishanth Menon <nm@ti.com>
(cherry picked from commit abe0852bf7911bc13f1c25951ab62a0a172015ec)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>