UPSTREAM: arm64: dts: ti: k3-j722s-main: Switch to 64-bit address space for PCIe0
commit 0fde00328cf8245bf498349a696ba22d6a4a1916 upstream.

The PCIe0 instance of PCIe in J722S SoC supports:
1. 128 MB address region in the 32-bit address space
2. 4 GB address region in the 64-bit address space

The default configuration is that of a 128 MB address region in the
32-bit address space. While this might be sufficient for most use-cases,
it is insufficient for supporting use-cases which require larger address
spaces. Therefore, switch to using the 64-bit address space with a 4 GB
address region.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Link: https://lore.kernel.org/r/20250422120042.3746004-7-s-vadapalli@ti.com
Signed-off-by: Nishanth Menon <nm@ti.com>
(cherry picked from commit e4281149e38703a34ab4d7682df021e297f4882a)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>