ANDROID: MEMORY: mm/page_alloc.c the pointer is not initialized
the pointer is not initialized in __alloc_pages_noprof.
And it will lead to panic when CONFIG_INIT_STACK_ALL_ZERO=n.
Because in the function flow, the pointer of the page is not
initialized, it becomes non-null, and the vendorhook may not
be executed. This leads to the non-null check being passed,
causing the out process to be skipped and the program continuing
with the subsequent steps.

Bug: 420376550
Change-Id: Ibd2eb2bd8a24e431ce90781bed3b72694c06bb42
Signed-off-by: Dezhen Wang <dezhen.wang@amlogic.com>