PENDING: pmdomain: ti_sci: Add LPM abort sequence to suspend path
Create ->suspend_late() and ->suspend_noirq() hooks to add abort sequence
to any driver within the PM domain with either of those hooks. If a driver
fails to suspend with those hooks, add a call to the DM to abort entering
the LPM. The suspend hooks of the PM domains driver inserts itself into
the suspend path of all devices connected to the TI SCI PM domain. So if
any device in the PM domain with either a ->suspend_late() or
->suspend_noirq hook fails to suspend, the PM domain drivers suspend hook
will send the abort call to the DM.

Adding an abort call in the ->suspend() hook is not necessary. TI SCI
suspend sends the message to the DM to prepare to enter a low power mode.
TI SCI suspend ALWAYS occurs after the ->suspend() hook for all TI SCI
devices has been called.

Signed-off-by: Kendall Willis <k-willis@ti.com>
(cherry picked from commit c41361fb83aa979f49eb5f4aa3f611bb56af37ff)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>