FROMLIST: drm/bridge: cdns-dsi: Remove broken fifo emptying check
The driver checks if "DPI(HFP) > DSI(HSS+HSA+HSE+HBP)", and rejects the
mode if not.

However, testing shows that this doesn't hold at all. I can set the hfp
to very small values, with no errors. The feedback from the HW team also
was that the check is not right, although it's not clear if there's a
way to validate the FIFO emptying.

The check rejects quite a lot of modes, apparently for no good reason,
so drop the check.

Link: https://lore.kernel.org/all/20250618-cdns-dsi-impro-v4-8-862c841dbe02@ideasonboard.com/
Tested-by: Parth Pancholi <parth.pancholi@toradex.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Harikrishna Shenoy <h-shenoy@ti.com>
(cherry picked from commit d21be51bf4c4d6c194825e072ed5b26e21082be1)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>