PENDING: net: ethernet: ti: cpsw-proxy-client: Use "port_id" to set Destination Tag
Starting with EthFw version 11.01.00.03, the "port_id" assigned to
Virtual MAC Ports corresponds to the Physical MAC Port. This allows the
Clients to send directed packets by setting the "Destination Tag" in the
DMA Descriptor to that of the Physical MAC Port. This is the intended
usage for Virtual MAC Ports. For Virtual Switch Ports, since the ALE
determines the destination Port(s) after a lookup, the "Destination Tag"
should be set to 0.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
(cherry picked from commit 617472159edbc6c7ca31e51a66ee9a27917b5076)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>