arm64: dts: k3-j722s: do not declare reset gpio for mf101hie panel
The ilitek ili251x controller driver does not require the reset gpio to
work properly.
However, if one is declared and the driver cannot acquire it, it will
error out and fail to probe, leading to the touchscreen being non
fonctionnal on the am67a-evm.
At this time, simply remove the reset gpio declaration as a quick fix
until a better fix can be implemented with the reset gpio.

Signed-off-by: Maxime Fournier <mfournier@baylibre.com>