PENDING: drivers: crypto: sa2ul: Program the Hash key for GCM
Hash key (AES-ECB(key,0)) is used for Tag calculation and it needs to be
done as part of setkey operation when we get the key. Populate the
buffer for hash key using standard helpers if we are going for GCM mode.

Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
(cherry picked from commit 12d44183cef555e669c3bf52718368b5d1e1f37b)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>